Elevate coordinate classes to layout_provider
